Move DrawShapeToolManager into its own unit
