+++ lib/TbUiLib/src/DrawShapeToolExtension.cpp
@@ -23,16 +23,8 @@
 #include <QPushButton>
 
 #include "mdl/Map.h"
-#include "mdl/Map_Nodes.h"
-#include "mdl/Transaction.h"
-#include "ui/DrawShapeToolExtensions.h"
 #include "ui/MapDocument.h"
 #include "ui/ViewConstants.h"
-
-#include "kd/contracts.h"
-#include "kd/ranges/to.h"
-
-#include <ranges>
 
 namespace tb::ui
 {
@@ -188,59 +180,5 @@ DrawShapeToolExtension::DrawShapeToolExtension(MapDocument& document)
 }
 
 DrawShapeToolExtension::~DrawShapeToolExtension() = default;
-
-DrawShapeToolExtensionManager::DrawShapeToolExtensionManager(MapDocument& document)
-  : m_extensions{createDrawShapeToolExtensions(document)}
-{
-  contract_pre(!m_extensions.empty());
-}
-
-const std::vector<DrawShapeToolExtension*> DrawShapeToolExtensionManager::extensions()
-  const
-{
-  return m_extensions
-         | std::views::transform([](const auto& extension) { return extension.get(); })
-         | kdl::ranges::to<std::vector>();
-}
-
-const DrawShapeToolExtension& DrawShapeToolExtensionManager::currentExtension() const
-{
-  return *m_extensions[m_currentExtensionIndex];
-}
-
-bool DrawShapeToolExtensionManager::setCurrentExtensionIndex(size_t currentExtensionIndex)
-{
-  if (currentExtensionIndex != m_currentExtensionIndex)
-  {
-    m_currentExtensionIndex = currentExtensionIndex;
-    currentExtensionDidChangeNotifier(m_currentExtensionIndex);
-    return true;
-  }
-
-  return false;
-}
-
-std::vector<DrawShapeToolExtensionPage*> DrawShapeToolExtensionManager::createToolPages(
-  QWidget* parent)
-{
-  auto toolPages = m_extensions | std::views::transform([&](const auto& extension) {
-                     auto* toolPage = extension->createToolPage(m_parameters, parent);
-                     m_notifierConnection +=
-                       toolPage->applyParametersNotifier.connect(applyParametersNotifier);
-                     return toolPage;
-                   })
-                   | kdl::ranges::to<std::vector>();
-
-  // update all tool pages to reflect the current parameter values
-  m_parameters.parametersDidChangeNotifier();
-
-  return toolPages;
-}
-
-Result<std::vector<mdl::Brush>> DrawShapeToolExtensionManager::createBrushes(
-  const vm::bbox3d& bounds) const
-{
-  return currentExtension().createBrushes(bounds, m_parameters);
-}
 
 } // namespace tb::ui
